refactor: introduce `KafkaConnectionConfig`
